feat(profile): AVO2-437 save profile info to database
