+++ src/settings/components/Account.tsx
@@ -1,4 +1,5 @@
-import React, { FunctionComponent, useState } from 'react';
+import { get } from 'lodash-es';
+import React, { FunctionComponent } from 'react';
 import { RouteComponentProps, withRouter } from 'react-router';
 import { compose } from 'redux';
 
@@ -12,32 +13,21 @@ import {
 	Grid,
 	Heading,
 	Spacer,
-	TextInput,
 } from '@viaa/avo2-components';
 import { Avo } from '@viaa/avo2-types';
 
-import { getProfileStamboekNumber } from '../../authentication/helpers/get-profile-info';
 import { redirectToServerSmartschoolLogin } from '../../authentication/helpers/redirects';
 import withUser from '../../shared/hocs/withUser';
-import toastService from '../../shared/services/toast-service';
 
 export interface AccountProps extends RouteComponentProps {
 	user?: Avo.User.User;
 }
 
-const Account: FunctionComponent<AccountProps> = ({ location, user }) => {
-	const [stamboekNumber, setStamboekNumber] = useState<string>(
-		getProfileStamboekNumber(user) || ''
-	);
-
+const Account: FunctionComponent<AccountProps> = ({ location, user, ...props }) => {
 	const getSsumAccountEditPage = () => {
 		// TODO replace this with a call to a proxy server route that forwards to the ssum page
 		// with the user already logged in and a redirect url back to this webpage after the user saves their changes
 		return 'https://account.hetarchief.be/';
-	};
-
-	const saveStamboekNumber = () => {
-		toastService.info(`Nog niet geimplementeerd: ${stamboekNumber}`);
 	};
 
 	return (
@@ -51,14 +41,8 @@ const Account: FunctionComponent<AccountProps> = ({ location, user }) => {
 									<Form type="standard">
 										<Heading type="h3">Account</Heading>
 										<FormGroup label="Email">
-											<span>test@testers.be</span>
+											<span>{get(user, 'mail')}</span>
 										</FormGroup>
-										{/*<FormGroup label="Wachtwoord">*/}
-										{/*	<span>123456</span>*/}
-										{/*</FormGroup>*/}
-										{/*<FormGroup label="Geldigheid">*/}
-										{/*	<span>Jouw account is nog 233 dagen geldig.</span>*/}
-										{/*</FormGroup>*/}
 										<Spacer margin="top-large">
 											<Alert type="info">
 												<span>
@@ -74,21 +58,6 @@ const Account: FunctionComponent<AccountProps> = ({ location, user }) => {
 											</Alert>
 										</Spacer>
 									</Form>
-
-									<div className="c-hr" />
-
-									<FormGroup label="Stamboeknummer / Lerarenkaart nummer" labelFor="stamboekNumber">
-										<TextInput
-											placeholder="00000000000-000000"
-											value={stamboekNumber}
-											onChange={setStamboekNumber}
-										/>
-									</FormGroup>
-									<Button
-										label="Stamboek nummer opslaan"
-										type="primary"
-										onClick={saveStamboekNumber}
-									/>
 
 									<div className="c-hr" />
